refactor(search): extract docs/code RAG tools to _RagToolsMixin
